requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-genai
weasyprint
jinja2
